smartcontract/neochat/requirements.txt
autopep8==1.3.3
bytecode==0.5
neo-boa==0.4.3
pycodestyle==2.3.1